Break post-hoist cycles structurally: fray.current_client, write_levanter_cache lives in levanter
Two real cycles surfaced in CI from the prior hoist sweep, fixed by
moving the integration glue into modules that own it:

1. **fray**: `fray/client.py` was hoisting `from fray.iris_backend import
   FrayIrisClient` to module top, but `fray.iris_backend` imports
   `JobAlreadyExists` from `fray.client` — partial-init ImportError on
   any `from fray import ...`. Extract `current_client()`,
   `set_current_client()`, and the contextvar into a new
   `fray.current_client` module. `fray.client` is now pure protocols +
   exceptions; backends import from it without cycling. `fray/__init__`
   re-exports `current_client`/`set_current_client` for compatibility.
   Update direct importers in `marin.execution.{executor,step_runner}`,
   `zephyr.execution`, `experiments/grug/dispatch.py`, and the
   `test_current_client.py` mocks (which now patch `fray.current_client`
   names, not the source modules).

2. **zephyr ↔ levanter**: `zephyr.writers` had hoisted
   `from levanter.store.cache import CacheMetadata, SerialCacheWriter`,
   cycling through `levanter.store.cache → zephyr.{Dataset, ZephyrContext}`.
   Move `write_levanter_cache` (a writer that writes levanter's cache
   format using levanter primitives) into `levanter.store.cache` where it
   belongs. Strip levanter-cache awareness from zephyr entirely:

   - delete `Dataset.write_levanter_cache` method, `WriteOp` levanter
     fields, the `"levanter_cache"` writer_type dispatch in `plan.py`,
     and the `write_levanter_cache` re-export from `zephyr/__init__.py`.
   - tokenize.py replaces `.write_levanter_cache(...)` with a
     `.map_shard(...)` closure that calls `levanter.store.cache.write_levanter_cache`
     directly, preserving distributed write semantics.
   - move `test_write_levanter_cache_end_to_end` from
     `lib/zephyr/tests/test_writers.py` into `lib/levanter/tests/test_new_cache.py`.

zephyr no longer references levanter at all in source.
