+++ experiments/grug/dispatch.py
@@ -8,8 +8,8 @@ import re
 from collections.abc import Callable
 from typing import TypeVar
 
-from fray.client import current_client
 from fray.cluster import ResourceConfig
+from fray.current_client import current_client
 from fray.types import Entrypoint, GpuConfig, JobRequest, TpuConfig, create_environment
 
 logger = logging.getLogger(__name__)
