+++ lib/fray/src/fray/client.py
@@ -5,18 +5,12 @@
 
 from __future__ import annotations
 
-import contextlib
-import contextvars
 import logging
 import time
-from collections.abc import Generator, Sequence
+from collections.abc import Sequence
 from typing import Any, Protocol
 
-from iris.client.client import get_iris_ctx
-
 from fray.actor import ActorGroup, ActorHandle, HostedActor
-from fray.iris_backend import FrayIrisClient
-from fray.local_backend import LocalClient
 from fray.types import ActorConfig, JobRequest, JobStatus, ResourceConfig
 
 logger = logging.getLogger(__name__)
@@ -156,39 +150,3 @@ def wait_all(
             sleep_secs = min(sleep_secs * 1.5, max_sleep_secs)
 
     return results  # type: ignore[return-value]
-
-
-_current_client_var: contextvars.ContextVar[Client | None] = contextvars.ContextVar("_current_client_var", default=None)
-
-
-def current_client() -> Client:
-    """Return the current fray Client.
-
-    Resolution order:
-        1. Explicitly set client (via set_current_client)
-        2. Auto-detect Iris environment (get_iris_ctx() returns context)
-        3. LocalClient() default
-    """
-
-    client = _current_client_var.get()
-    if client is not None:
-        logger.info("current_client: using explicitly set client")
-        return client
-
-    ctx = get_iris_ctx()
-    if ctx is not None:
-        logger.info("current_client: using Iris backend (auto-detected)")
-        return FrayIrisClient.from_iris_client(ctx.client)
-
-    logger.info("current_client: using LocalClient (fallback)")
-    return LocalClient()
-
-
-@contextlib.contextmanager
-def set_current_client(client: Client) -> Generator[Client, None, None]:
-    """Context manager that sets the current client and restores on exit."""
-    token = _current_client_var.set(client)
-    try:
-        yield client
-    finally:
-        _current_client_var.reset(token)
