+++ lib/fray/tests/test_current_client.py
@@ -6,13 +6,13 @@
 from unittest.mock import MagicMock, patch
 
 import pytest
-from fray.client import current_client, set_current_client
+from fray.current_client import current_client, set_current_client
 from fray.local_backend import LocalClient
 
 
 def test_default_returns_local_client():
     """When no context is set, should return LocalClient."""
-    with patch("iris.client.client.get_iris_ctx", return_value=None):
+    with patch("fray.current_client.get_iris_ctx", return_value=None):
         client = current_client()
         assert isinstance(client, LocalClient)
 
@@ -42,8 +42,8 @@ def test_iris_auto_detection_with_context():
     mock_iris_client_lib = MagicMock()
     mock_ctx.client = mock_iris_client_lib
 
-    with patch("iris.client.client.get_iris_ctx", return_value=mock_ctx):
-        with patch("fray.iris_backend.FrayIrisClient") as mock_client_cls:
+    with patch("fray.current_client.get_iris_ctx", return_value=mock_ctx):
+        with patch("fray.current_client.FrayIrisClient") as mock_client_cls:
             mock_fray_client = MagicMock()
             mock_client_cls.from_iris_client.return_value = mock_fray_client
 
@@ -58,8 +58,8 @@ def test_iris_auto_detection_reuses_client():
     mock_iris_client_lib = MagicMock()
     mock_ctx.client = mock_iris_client_lib
 
-    with patch("iris.client.client.get_iris_ctx", return_value=mock_ctx):
-        with patch("fray.iris_backend.FrayIrisClient") as mock_client_cls:
+    with patch("fray.current_client.get_iris_ctx", return_value=mock_ctx):
+        with patch("fray.current_client.FrayIrisClient") as mock_client_cls:
             mock_fray_client = MagicMock()
             mock_client_cls.from_iris_client.return_value = mock_fray_client
 
@@ -70,7 +70,7 @@ def test_iris_auto_detection_reuses_client():
 
 def test_iris_not_detected_when_no_context():
     """Should not detect Iris when get_iris_ctx() returns None."""
-    with patch("iris.client.client.get_iris_ctx", return_value=None):
+    with patch("fray.current_client.get_iris_ctx", return_value=None):
         client = current_client()
         assert isinstance(client, LocalClient)
 
@@ -82,7 +82,7 @@ def test_explicit_client_overrides_auto_detection():
     mock_ctx.client = mock_iris_client_lib
 
     explicit = LocalClient(max_threads=1)
-    with patch("iris.client.client.get_iris_ctx", return_value=mock_ctx):
+    with patch("fray.current_client.get_iris_ctx", return_value=mock_ctx):
         with set_current_client(explicit):
             # Should return explicit client, not auto-detected Iris client
             assert current_client() is explicit
