+++ lib/zephyr/src/zephyr/dataset.py
@@ -205,18 +205,16 @@ class WindowOp:
 class WriteOp:
     """Unified write operation for all output formats.
 
-    Supports writing to JSONL, Parquet, Levanter cache, or binary formats.
+    Supports writing to JSONL, Parquet, or binary formats.
     The writer_type determines which writer function is used.
     Supports path patterns with {shard}, {total}, {basename} substitutions,
     or a callable that takes (shard_idx, total_shards) and returns the output path.
     """
 
     output_pattern: Callable[[int, int], str]
-    writer_type: Literal["jsonl", "parquet", "levanter_cache", "binary", "vortex"]
+    writer_type: Literal["jsonl", "parquet", "binary", "vortex"]
 
     # Format-specific parameters (only used by relevant writer)
-    levanter_metadata: dict[str, Any] | None = None
-    levanter_batch_size: int | None = None
     schema: object | None = None  # For parquet (pyarrow.Schema)
     skip_existing: bool = False  # Skip writing if output file already exists
 
@@ -753,38 +751,6 @@ class Dataset(Generic[T]):
                     _normalize_output_pattern(output_pattern),
                     writer_type="vortex",
                     schema=schema,
-                    skip_existing=skip_existing,
-                ),
-            ],
-        )
-
-    def write_levanter_cache(
-        self,
-        output_pattern: str | Callable[[int, int], str],
-        metadata: dict[str, Any],
-        skip_existing: bool = False,
-        batch_size: int | None = None,
-    ) -> Dataset[str]:
-        """Write tokenized records to Levanter cache format.
-
-        Writes records to Levanter's TreeStore/JaggedArrayStore format for use
-        in training. Each shard creates a separate cache directory.
-        The output pattern supports substitutions: {shard:05d}, {total:05d}, {basename}
-        or can be a callable that takes (shard_idx, total_shards) and returns the output path.
-
-        Args:
-            batch_size: Number of records to accumulate before flushing to disk.
-                Defaults to 16384. Lower values reduce peak memory for large documents.
-        """
-        return Dataset(
-            self.source,
-            [
-                *self.operations,
-                WriteOp(
-                    _normalize_output_pattern(output_pattern),
-                    writer_type="levanter_cache",
-                    levanter_metadata=metadata,
-                    levanter_batch_size=batch_size,
                     skip_existing=skip_existing,
                 ),
             ],
